Introducing Immutable Objects
Another approach to replace Immutable Structs. For this first commit, I
am trying it for the base Model class. Similar to Immutable Struct,
Immutable Object is instantiated with its attribute values, and after
that it does not accept changes to any of its state.

The difference is the API offered. While Struct initializes a new class,
Immutable Object exposes an `attribute` macro/class method to
dynamically define attributes. This is arguably a nicer API for
developers, since it is inspired by Active Model's Attributes API.

The downside is that, while instances are immutable, the Immutable
Object class itself is not, since it holds a global state of attribute
names accepted that is mutated using the `attributes` macro.